feat: update | docker compose | dockerfile | env && add makefile

+++ srcs/requirements/wordpress/tools/init.sh
@@ -2,16 +2,19 @@
 set -e
 
 until mysqladmin -h mariadb -u${MYSQL_USER} -p${MYSQL_PASSWORD} ping --silent; do
+    echo "Waiting for MariaDB..."
     sleep 2
 done
 
 cd /var/www/html
 
-cp wp-config-sample.php wp-config.php
+if [ ! -f wp-config.php ]; then
+    cp wp-config-sample.php wp-config.php
 
-sed -i "s/database_name_here/${MYSQL_DATABASE}/" wp-config.php
-sed -i "s/username_here/${MYSQL_USER}/" wp-config.php
-sed -i "s/password_here/${MYSQL_PASSWORD}/" wp-config.php
-sed -i "s/localhost/mariadb/" wp-config.php
+    sed -i "s/database_name_here/${MYSQL_DATABASE}/" wp-config.php
+    sed -i "s/username_here/${MYSQL_USER}/" wp-config.php
+    sed -i "s/password_here/${MYSQL_PASSWORD}/" wp-config.php
+    sed -i "s/localhost/mariadb/" wp-config.php
+fi
 
 exec php-fpm8.2 -F
